+++ VpNet/trunk/VpNet/Math/MathHelper.cs
@@ -31,42 +31,42 @@ namespace VpNet
 {
     public static class MathHelper
     {
-        public const float E = (float)Math.E;
-        public const float Log10E = 0.4342945f;
-        public const float Log2E = 1.442695f;
-        public const float Pi = (float)Math.PI;
-        public const float PiOver2 = (float)(Math.PI / 2.0);
-        public const float PiOver4 = (float)(Math.PI / 4.0);
-        public const float TwoPi = (float)(Math.PI * 2.0);
+        public const double E = (double)Math.E;
+        public const double Log10E = 0.4342945f;
+        public const double Log2E = 1.442695f;
+        public const double Pi = (double)Math.PI;
+        public const double PiOver2 = (double)(Math.PI / 2.0);
+        public const double PiOver4 = (double)(Math.PI / 4.0);
+        public const double TwoPi = (double)(Math.PI * 2.0);
 
-        public static float Barycentric(float value1, float value2, float value3, float amount1, float amount2)
+        public static double Barycentric(double value1, double value2, double value3, double amount1, double amount2)
         {
             return value1 + (value2 - value1) * amount1 + (value3 - value1) * amount2;
         }
 
-        public static float CatmullRom(float value1, float value2, float value3, float value4, float amount)
+        public static double CatmullRom(double value1, double value2, double value3, double value4, double amount)
         {
             // Using formula from http://www.mvps.org/directx/articles/catmull/
             // Internally using doubles not to lose precission
             double amountSquared = amount * amount;
             double amountCubed = amountSquared * amount;
-            return (float)(0.5f * (2.0f * value2 +
+            return (double)(0.5f * (2.0f * value2 +
                 (value3 - value1) * amount +
                 (2.0f * value1 - 5.0f * value2 + 4.0f * value3 - value4) * amountSquared +
                 (3.0f * value2 - value1 - 3.0f * value3 + value4) * amountCubed));
         }
 
-        public static float Clamp(float value, float min, float max)
+        public static double Clamp(double value, double min, double max)
         {
             return value > max ? max : (value < min ? min : value);
         }
 
-        public static float Distance(float value1, float value2)
+        public static double Distance(double value1, double value2)
         {
             return Math.Abs(value1 - value2);
         }
 
-        public static float Hermite(float value1, float tangent1, float value2, float tangent2, float amount)
+        public static double Hermite(double value1, double tangent1, double value2, double tangent2, double amount)
         {
             // All transformed to double not to lose precission
             // Otherwise, for high numbers of param:amount the result is NaN instead of Infinity
@@ -83,55 +83,55 @@ namespace VpNet
                     (3.0f * v2 - 3.0f * v1 - 2.0f * t1 - t2) * sSquared +
                     t1 * s +
                     v1;
-            return (float)result;
+            return (double)result;
         }
 
 
-        public static float Lerp(float value1, float value2, float amount)
+        public static double Lerp(double value1, double value2, double amount)
         {
             return value1 + (value2 - value1) * amount;
         }
 
-        public static float Max(float value1, float value2)
+        public static double Max(double value1, double value2)
         {
             return Math.Max(value1, value2);
         }
 
-        public static float Min(float value1, float value2)
+        public static double Min(double value1, double value2)
         {
             return Math.Min(value1, value2);
         }
 
-        public static float SmoothStep(float value1, float value2, float amount)
+        public static double SmoothStep(double value1, double value2, double amount)
         {
             // It is expected that 0 < amount < 1
             // If amount < 0, return value1
             // If amount > 1, return value2
-            float result = MathHelper.Clamp(amount, 0f, 1f);
+            double result = MathHelper.Clamp(amount, 0f, 1f);
             result = MathHelper.Hermite(value1, 0f, value2, 0f, result);
             return result;
         }
 
-        public static float ToDegrees(float radians)
+        public static double ToDegrees(double radians)
         {
             // This method uses double precission internally,
-            // though it returns single float
+            // though it returns single double
             // Factor = 180 / pi
-            return (float)(radians * 57.295779513082320876798154814105);
+            return (double)(radians * 57.295779513082320876798154814105);
         }
 
-        public static float ToRadians(float degrees)
+        public static double ToRadians(double degrees)
         {
             // This method uses double precission internally,
-            // though it returns single float
+            // though it returns single double
             // Factor = pi / 180
-            return (float)(degrees * 0.017453292519943295769236907684886);
+            return (double)(degrees * 0.017453292519943295769236907684886);
         }
 
 
-        public static float WrapAngle(float angle)
+        public static double WrapAngle(double angle)
         {
-            angle = (float)Math.IEEERemainder((double)angle, 6.2831854820251465); //2xPi precission is double
+            angle = (double)Math.IEEERemainder((double)angle, 6.2831854820251465); //2xPi precission is double
             if (angle <= -3.141593f)
             {
                 angle += 6.283185f;
